Fix: show user in chatline
The server send connection info for messages when that applies, but
connections come and go, and they will probably not exist when a player
reconnects and gets messages from the event cache.

The text is also preceded with the player name. In the case of normal
messages that is in the form of "<leader name>". Freeciv-web still uses
direct html incorporation in a los of places, the chatbox being one of
them, so the browser saw a tag there and the name didn't appear.

This was "solved" for longturn (the type where this problem is most
visible) by adding the username to the message when sending it to the
server, making it part of the text. That's just a hack and the root
cause of the problem is still there, as shown with permanent games that
are not longturn.

This patch is still a hack, the real solution pending the removal of
server patches related to messages and the implementation of message
parsing and escaping in the client, but it goes in that direction.