Add option to use player colors
Player colors are already assigned by the server. The command
plrcolormode allows different assignment options (per order, random,
by team, by nation, manually), so the client can but doesn't need to
compute them. The command playercolor allows the admin to change the
color of a player even in game.

The current choice of using the color most used in the nation's flag is
nice, but it also leads to easy clashes between players and terrains.