Move compilation check out of the message loop
The check for the extraction of client_handle_packet can be done on
client initialization, instead of once for each message received. Also
make it visible, so that in the improbable case that it is fired on
final users, they see there has been an error instead of a non-working
pregame page.