fix: find tile for canvas coords
canvas_pos_to_tile does not return a tile if there is not one under the
given canvas coordinates. When looking for the next unit to focus we
want a tile to compare distances to even in that case, so use a modified
version that always returns a tile.

Fix #3